+++ src/lib.rs
@@ -2,6 +2,8 @@
 
 // `SPDX-License-Identifier: MIT OR Apache-2.0`
 
+/// Public API
+pub mod api;
 /// Implementations of the `rtlambda` API for different HTTP backends.
 pub mod backends;
 /// A collection of traits and default implementations for them, representing the library's core data structures.
@@ -19,8 +21,10 @@ pub static LAMBDA_VER: &str = "2018-06-01";
 /// A prelude that contains all the relevant imports when using the library's default runtime implementation,
 /// which currently ships with a [ureq](https://crates.io/crates/ureq) based HTTP Backend and [serde_json](https://crates.io/crates/serde_json) for serialization.
 pub mod prelude {
+    pub use crate::api::LambdaContext;
+    pub use crate::api::*;
     pub use crate::backends::ureq::*;
-    pub use crate::data::context::{EventContext, LambdaContext};
+    pub use crate::data::context::EventContext;
     pub use crate::runtime::event_handler::EventHandler;
     pub use crate::runtime::{DefaultRuntime, LambdaRuntime};
     pub use crate::LAMBDA_VER;
